styling bootstrap

+++ src/Controller/PostController.php
@@ -19,17 +19,10 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 class PostController extends AbstractController
 {
     protected $session;
-    public function __construct(SessionInterface $session){
-
+    public function __construct(SessionInterface $session)
+    {
         // must be logged in to view this controller
         $this->session = $session;
-        $is_logged_in  = $this->session->get('is_logged_in');
-
-        if($is_logged_in != true){
-            return $this->render('security/login.html.twig', [
-            ]);
-        }
-       // parent::__construct();
     }
 
     /**
@@ -37,9 +30,11 @@ class PostController extends AbstractController
      */
     public function index(PostRepository $postRepository): Response
     {
-        return $this->render('post/index.html.twig', [
-            'posts' => $postRepository->findAll(),
-        ]);
+        if($this->session->get('is_logged_in') == true){
+            return $this->render('post/index.html.twig', ['posts' => $postRepository->findAll()]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 
     /**
@@ -47,26 +42,30 @@ class PostController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $post = new Post();
-        $form = $this->createForm(PostType::class, $post);
-        $form->handleRequest($request);
+        if($this->session->get('is_logged_in') == true){
+            $post = new Post();
+            $form = $this->createForm(PostType::class, $post);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
 
-            // set user id
-            $post->setUserId(intval($this->session->get('user_id')));
+                // set user id
+                $post->setUserId(intval($this->session->get('user_id')));
 
-            $entityManager->persist($post);
-            $entityManager->flush();
+                $entityManager->persist($post);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('post_index');
+                return $this->redirectToRoute('post_index');
+            }
+
+            return $this->render('post/new.html.twig', [
+                'post' => $post,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('post/new.html.twig', [
-            'post' => $post,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -74,9 +73,13 @@ class PostController extends AbstractController
      */
     public function show(Post $post): Response
     {
-        return $this->render('post/show.html.twig', [
-            'post' => $post,
-        ]);
+        if($this->session->get('is_logged_in') == true) {
+            return $this->render('post/show.html.twig', [
+                'post' => $post,
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 
     /**
@@ -84,20 +87,23 @@ class PostController extends AbstractController
      */
     public function edit(Request $request, Post $post): Response
     {
-        $form = $this->createForm(PostType::class, $post);
-        //$form->remove('user_id');
-        $form->handleRequest($request);
+        if($this->session->get('is_logged_in') == true) {
+            $form = $this->createForm(PostType::class, $post);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('post_index');
+                return $this->redirectToRoute('post_index');
+            }
+
+            return $this->render('post/edit.html.twig', [
+                'post' => $post,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('post/edit.html.twig', [
-            'post' => $post,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -105,12 +111,18 @@ class PostController extends AbstractController
      */
     public function delete(Request $request, Post $post): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($post);
-            $entityManager->flush();
-        }
+        if($this->session->get('is_logged_in') == true)
+        {
+            if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token')))
+            {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($post);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('post_index');
+            return $this->redirectToRoute('post_index');
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 }
