+++ src/Controller/UserController.php
@@ -29,10 +29,14 @@ class UserController extends AbstractController
      * @Route("/", name="user_index", methods={"GET"})
      */
     public function index(UserRepository $userRepository): Response
-    {//echo '<pre>';print_r($userRepository->findAll()); exit();
-        return $this->render('user/index.html.twig', [
-            'users' => $userRepository->findAll(),
-        ]);
+    {
+        if($this->session->get('is_logged_in') == true){
+            return $this->render('user/index.html.twig', [
+                'users' => $userRepository->findAll(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 
     /**
@@ -48,6 +52,7 @@ class UserController extends AbstractController
             $entityManager = $this->getDoctrine()->getManager();
             $plainPassword = $user->getPassword();
             $encoded       = $this->passwordEncoder->encodePassword($user, $plainPassword);
+
             $user->setPassword($encoded);
 
             $entityManager->persist($user);
@@ -75,9 +80,14 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
-        return $this->render('user/show.html.twig', [
-            'user' => $user,
-        ]);
+        if($this->session->get('is_logged_in') == true) {
+            return $this->render('user/show.html.twig', [
+                'user' => $user,
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
+
     }
 
     /**
@@ -85,27 +95,31 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user): Response
     {
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
+        if($this->session->get('is_logged_in') == true) {
+            $form = $this->createForm(UserType::class, $user);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            //$this->getDoctrine()->getManager()->flush();
-            $entityManager = $this->getDoctrine()->getManager();
+            if ($form->isSubmitted() && $form->isValid()) {
+                //$this->getDoctrine()->getManager()->flush();
+                $entityManager = $this->getDoctrine()->getManager();
 
-            $plainPassword = $user->getPassword();
-            $encoded = $this->passwordEncoder->encodePassword($user, $plainPassword);
-            $user->setPassword($encoded);
+                $plainPassword = $user->getPassword();
+                $encoded = $this->passwordEncoder->encodePassword($user, $plainPassword);
+                $user->setPassword($encoded);
 
-            $entityManager->persist($user);
-            $entityManager->flush();
+                $entityManager->persist($user);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('user_index');
+                return $this->redirectToRoute('user_index');
+            }
+
+            return $this->render('user/edit.html.twig', [
+                'user' => $user,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('user/edit.html.twig', [
-            'user' => $user,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -113,12 +127,17 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
-        }
+        if($this->session->get('is_logged_in') == true) {
+            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token')))
+            {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($user);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('user_index');
+            return $this->redirectToRoute('user_index');
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 }
